+++ driver/web/docs/gen/gen_types.go
@@ -232,15 +232,16 @@ type APIKey struct {
 
 // Account defines model for Account.
 type Account struct {
-	AppOrg      *ApplicationOrganization `json:"app_org,omitempty"`
-	AuthTypes   *[]AccountAuthType       `json:"auth_types,omitempty"`
-	Devices     *[]Device                `json:"devices,omitempty"`
-	Fields      *AccountFields           `json:"fields,omitempty"`
-	Groups      *[]AppOrgGroup           `json:"groups,omitempty"`
-	Permissions *[]Permission            `json:"permissions,omitempty"`
-	Preferences *map[string]interface{}  `json:"preferences,omitempty"`
-	Profile     *Profile                 `json:"profile,omitempty"`
-	Roles       *[]AppOrgRole            `json:"roles,omitempty"`
+	AppOrg        *ApplicationOrganization `json:"app_org,omitempty"`
+	AuthTypes     *[]AccountAuthType       `json:"auth_types,omitempty"`
+	Devices       *[]Device                `json:"devices,omitempty"`
+	Fields        *AccountFields           `json:"fields,omitempty"`
+	Groups        *[]AppOrgGroup           `json:"groups,omitempty"`
+	Permissions   *[]Permission            `json:"permissions,omitempty"`
+	Preferences   *map[string]interface{}  `json:"preferences"`
+	Profile       *Profile                 `json:"profile,omitempty"`
+	Roles         *[]AppOrgRole            `json:"roles,omitempty"`
+	SystemConfigs *map[string]interface{}  `json:"system_configs"`
 }
 
 // AccountAuthType defines model for AccountAuthType.
@@ -509,6 +510,7 @@ type PartialAccount struct {
 	Params         *map[string]interface{} `json:"params"`
 	Permissions    []Permission            `json:"permissions"`
 	Roles          []AppOrgRole            `json:"roles"`
+	SystemConfigs  *map[string]interface{} `json:"system_configs"`
 }
 
 // Permission defines model for Permission.
@@ -996,6 +998,7 @@ type SharedResAccount struct {
 	Preferences    *map[string]interface{}  `json:"preferences"`
 	Profile        *ProfileFields           `json:"profile,omitempty"`
 	Roles          *[]AppOrgRole            `json:"roles,omitempty"`
+	SystemConfigs  *map[string]interface{}  `json:"system_configs"`
 }
 
 // SharedResAccountCheck defines model for _shared_res_AccountCheck.
@@ -1250,6 +1253,9 @@ type DeleteAdminApplicationAccountsIdRolesJSONBody AdminReqRevokeRolesFromAccoun
 
 // PutAdminApplicationAccountsIdRolesJSONBody defines parameters for PutAdminApplicationAccountsIdRoles.
 type PutAdminApplicationAccountsIdRolesJSONBody AdminReqGrantRolesToAccount
+
+// PutAdminApplicationAccountsIdSystemConfigsJSONBody defines parameters for PutAdminApplicationAccountsIdSystemConfigs.
+type PutAdminApplicationAccountsIdSystemConfigsJSONBody map[string]interface{}
 
 // PostAdminApplicationGroupsJSONBody defines parameters for PostAdminApplicationGroups.
 type PostAdminApplicationGroupsJSONBody AdminReqApplicationGroup
@@ -1701,6 +1707,9 @@ type DeleteAdminApplicationAccountsIdRolesJSONRequestBody DeleteAdminApplication
 
 // PutAdminApplicationAccountsIdRolesJSONRequestBody defines body for PutAdminApplicationAccountsIdRoles for application/json ContentType.
 type PutAdminApplicationAccountsIdRolesJSONRequestBody PutAdminApplicationAccountsIdRolesJSONBody
+
+// PutAdminApplicationAccountsIdSystemConfigsJSONRequestBody defines body for PutAdminApplicationAccountsIdSystemConfigs for application/json ContentType.
+type PutAdminApplicationAccountsIdSystemConfigsJSONRequestBody PutAdminApplicationAccountsIdSystemConfigsJSONBody
 
 // PostAdminApplicationGroupsJSONRequestBody defines body for PostAdminApplicationGroups for application/json ContentType.
 type PostAdminApplicationGroupsJSONRequestBody PostAdminApplicationGroupsJSONBody